chore: move ui-test-utils.ts
