+++ src/components/pagination/pagination-popover.ui.spec.tsx
@@ -1,4 +1,4 @@
-import { vi, expect, test } from "#src/ui-test-utils/index.ts";
+import { vi, expect, test } from "#utils/ui-test-utils.ts";
 import { render, waitFor } from "@testing-library/preact";
 import { PaginationMenuAndPopover } from "./pagination-popover.tsx";
